+++ npm-api-maker/src/super-admin/layout/index.jsx
@@ -11,7 +11,7 @@ import PropTypes from "prop-types"
 import PropTypesExact from "prop-types-exact"
 import {shapeComponent} from "set-state-compare/build/shape-component.js"
 import Text from "../../utils/text"
-import useBreakpoint from "../../use-breakpoint.js"
+import {useBreakpoint} from "responsive-breakpoints"
 import useCurrentUser from "../../use-current-user.js"
 import useI18n from "i18n-on-steroids/src/use-i18n.mjs"