+++ npm-api-maker/src/use-validation-errors.js
@@ -1,6 +1,6 @@
 import events from "./events.js"
 import {useCallback} from "react"
-import useEventEmitter from "./use-event-emitter.js"
+import useEventEmitter from "ya-use-event-emitter"
 import useShape from "set-state-compare/build/use-shape.js"
 
 /**