+++ ruby-gem/spec/dummy/app/javascript/routes/sessions/new.jsx
@@ -5,7 +5,7 @@ import Devise from "@kaspernj/api-maker/build/devise.js"
 import {FlashNotifications} from "flash-notifications"
 import Input from "@kaspernj/api-maker/build/bootstrap/input"
 import Layout from "components/layout"
-import useEventEmitter from "@kaspernj/api-maker/build/use-event-emitter.js"
+import useEventEmitter from "ya-use-event-emitter"
 
 export default memo(shapeComponent(class SessionsNew extends ShapeComponent {
   setup() {